+++ SeleniumHelpers.cs
@@ -1,11 +1,20 @@
 ﻿using OpenQA.Selenium.Support.UI;
 using static TimesheetTimothy.Program;
 using static SeleniumExtras.WaitHelpers.ExpectedConditions;
+using WebDriverManager.DriverConfigs.Impl;
+using WebDriverManager;
 
 namespace TimesheetTimothy;
 
 public static class SeleniumHelpers
 {
+    public static void SetUpChromeDriver()
+    {
+        new DriverManager().SetUpDriver(new ChromeConfig());
+
+        Driver.Navigate().GoToUrl("https://timesheets.dialoggroup.biz/?company=accesstesting");
+    }
+
     public static void SendKeys(By selector, string input)
     {
         Driver.FindElement(selector).SendKeys(input);
